requirements.txt
snowflake-connector-python
google-genai 
python-dotenv
streamlit
pandas